src/scheduler/requirements.txt
confluent-kafka==1.7.0
fastapi==0.104.1
python-dotenv==0.19.1
uvicorn==0.15.0
pydantic==1.8.2